refactor: rework blocks registrar
